drivers: usb: mtu3: Fix status handling of zero length setup packet
UDC drivers are not supposed to rely on USB_GADGET_DELAYED_STATUS. The
composite framework always sent out the ZLP packets, so autocompletion
was only required because the MTU3 driver internally relied on it.
Remove the status autocompletion, always assume status is delayed and
convert delayed_status to a state (because it really is a state and not
just a flag).

Change-Id: Iea61c55e412d2c0474fd2a1c49ee453dbe1e94b5